HW06: renaming classes and interfaces after review;
