+++ hw06-atm/src/main/java/ru/otus/atm/AtmService.java
@@ -2,7 +2,7 @@ package ru.otus.atm;
 
 import java.util.Map;
 
-public interface IAtmService {
+public interface AtmService {
 
     long serviceLogin(long serviceKey);
