+++ hw06-atm/src/main/java/ru/otus/atm/AtmUser.java
@@ -3,7 +3,7 @@ package ru.otus.atm;
 import java.util.List;
 import java.util.Map;
 
-public interface IAtmUser {
+public interface AtmUser {
 
     long userLogin(long cardNumber, int pin);
