--- hw06-atm/src/main/java/ru/otus/atm/ICells.java
@@ -1,18 +1,0 @@
-package ru.otus.atm;
-
-import java.util.List;
-import java.util.Map;
-
-interface ICells {
-
-    boolean inputBanknotes(Map<Nominal, Integer> banknotes);
-
-    Map<Nominal, Integer> takeAmountOut(int sum);
-
-    List<Nominal> getAvailableNominals();
-
-    boolean setState (Map<Nominal, Integer> cellsState);
-
-    Map<Nominal, Integer> getState ();
-
-}
